requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepsearch-glm==1.0.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docling==2.15.1
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==3.4.0
docutils==0.21.2
docx2txt==0.8
duckduckgo_search==6.3.3
durationpy==0.10
easyocr==1.7.2
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.9.0
filelock==3.20.1
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.2.0
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.5
langchain-huggingface==0.1.0
langchain-ollama==1.0.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.1
latex2mathml==3.78.1
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.14.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai==1.59.8
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pillow==12.0.0
posthog==5.4.0
primp==0.15.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pyclipper==1.4.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.24.12
PyMySQL==1.1.1
pypdf==6.5.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.2.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
streamlit==1.52.2
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.1.2
textblob==0.18.0.post0
threadpoolctl==3.6.0
tifffile==2025.12.20
tiktoken==0.12.0
tokenizers==0.21.4
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.48.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.5.1
typer==0.12.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wikipedia==1.4.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==0.6.2
zipp==3.23.0
zstandard==0.25.0
